#329 Moved TIC collector to a lib
